Add storage and device property getter AOTI shim functions to the header-only common_shims_slim library:

1. `aoti_torch_get_storage_offset()` - Returns the storage offset (SlimTensor: real offset, ETensor: always 0)
2. `aoti_torch_get_storage_size()` - Returns storage size in bytes
3. `aoti_torch_get_device_type()` - Returns device type (SlimTensor: real type, ETensor: CPU=0)
4. `aoti_torch_get_device_index()` - Returns device index (SlimTensor: real index, ETensor: 0)

Differential Revision: [D90126251](https://our.internmc.facebook.com/intern/diff/D90126251/)

[ghstack-poisoned]

+++ kernels/portable/cpu/vec_ops.h
@@ -179,7 +179,7 @@ template <typename T>
 inline float vec_powerf(const T* x, size_t size) {
   float sum = 0;
   for (const auto i : c10::irange(size)) {
-    sum += x[i] * x[i];
+    sum += static_cast<float>(x[i]) * x[i];
   }
   return sum;
 }
